TST: split up tests into single files
